Move classes to structure namespace
