+++ tests/back/Acceptance/Structure/Attribute/Group/InMemoryAttributeGroupRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Test\Acceptance\AttributeGroup;
+namespace Akeneo\Test\Acceptance\Structure\Attribute\Group;
 
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
 use Akeneo\Test\Acceptance\Common\NotImplementedException;
