+++ tests/back/Acceptance/Structure/Attribute/InMemoryAttributeRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Test\Acceptance\Attribute;
+namespace Akeneo\Test\Acceptance\Structure\Attribute;
 
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
 use Akeneo\Test\Acceptance\Common\NotImplementedException;
