+++ tests/back/Acceptance/Structure/Attribute/Option/InMemoryAttributeOptionRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Test\Acceptance\AttributeOption;
+namespace Akeneo\Test\Acceptance\Structure\Attribute\Option;
 
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
 use Akeneo\Test\Acceptance\Common\NotImplementedException;
