+++ tests/back/Acceptance/Structure/Family/InMemoryFamilyRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Test\Acceptance\Family;
+namespace Akeneo\Test\Acceptance\Structure\Family;
 
 use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
