+++ tests/back/Acceptance/Structure/Family/Variant/InMemoryFamilyVariantRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Test\Acceptance\FamilyVariant;
+namespace Akeneo\Test\Acceptance\Structure\Family\Variant;
 
 use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
